restructured the code to follow clean architecture standards
