+++ go.mod
@@ -5,7 +5,6 @@ go 1.16
 require (
 	github.com/gin-gonic/gin v1.7.2
 	github.com/go-sql-driver/mysql v1.6.0
-	github.com/mitchellh/gox v1.0.1 // indirect
 	github.com/privatesquare/bkst-go-utils v1.1.0
 	github.com/spf13/viper v1.7.1
 )
